+++ greentest/test__issue639.py
@@ -1,0 +1,4 @@
+# Test idle
+import gevent
+gevent.sleep()
+gevent.idle()